Move and update generateAccounts.js tool.
